+++ functions/Get-TibberProduction.ps1
@@ -9,7 +9,7 @@
         $maxProd = $response | Sort-Object -Property production -Descending | Select-Object -First 1
         Write-Host "Max power production $($maxProd.profit) $($maxProd.currency) ($($maxProd.production) $($maxProd.productionUnit) at $($maxProd.unitPrice)): $(([DateTime]$maxProd.from).ToString('HH:mm')) - $(([DateTime]$maxProd.to).ToString('HH:mm on yyyy-MM-dd'))"
     .Link
-        Invoke-TibberGraphQLQuery
+        Invoke-TibberQuery
     .Link
         https://developer.tibber.com/docs/reference#production
     #>
@@ -47,7 +47,7 @@
     )
 
     dynamicParam {
-        $dynamicParameters = Invoke-TibberGraphQLQuery -DynamicParameter
+        $dynamicParameters = Invoke-TibberQuery -DynamicParameter
         return $dynamicParameters
     }
 
@@ -66,7 +66,7 @@
             $homeNode = 'homes'
             $query += "$homeNode{ "
         }
-        $query += "production($arguments){ nodes{ $Fields, __typename } }"
+        $query += "production($arguments){ nodes{ $($Fields -join ','),__typename } }"
         $query += "}}}" # close query
 
         # Setup parameters
@@ -81,7 +81,7 @@
         } + $dynamicParametersValues
 
         # Call the GraphQL query API
-        $out = Invoke-TibberGraphQLQuery @splat
+        $out = Invoke-TibberQuery @splat
 
         # Output the object
         $out.viewer.$homeNode.production.nodes